base controller module and function for authenticating user

+++ controllers/todos_controller.go
@@ -8,29 +8,19 @@ import (
     "to-go/models"
     "to-go/storage"
     "github.com/gorilla/mux"
-    "strings"
 )
-
-type TodoRequest struct {
-    Todo models.Todo
-    AuthToken string
-}
-
-type TodoFailResponse struct {
-    Message    string `json:"message"`
-}
-
-func newTodoFailResponse() TodoFailResponse {
-    response := TodoFailResponse{}
-    response.Message = "Could not find a todo item with the given ID."
-    return response
-}
 
 func TodosIndex(w http.ResponseWriter, r *http.Request) {
     db := storage.GetActiveDB()
+    user := Authenticate(r, db)
+
+    if user.ID == 0 {
+        RespondWithMessage(w, "Unable to authenticate user.")
+        return
+    }
+
     todos := models.Todos{}
-    authToken := strings.Join(r.Header["X-Auth-Token"], "")
-    db.Joins("join users on user_id = users.id").Where("users.auth_token = ?", authToken).Find(&todos)
+    db.Where("user_id = ?", user.ID).Find(&todos)
 
     w.Header().Set("Content-Type", "application/json; charset=UTF-8")
     w.WriteHeader(http.StatusOK)
@@ -41,18 +31,20 @@ func TodosIndex(w http.ResponseWriter, r *http.Request) {
 
 func TodoShow(w http.ResponseWriter, r *http.Request) {
     db := storage.GetActiveDB()
+    user := Authenticate(r, db)
+
+    if user.ID == 0 {
+        RespondWithMessage(w, "Unable to authenticate user.")
+        return
+    }
+
     vars := mux.Vars(r)
     todoId := vars["todoId"]
     todo := models.Todo{}
-    db.Find(&todo, todoId)
+    db.Where("user_id = ?", user.ID).Find(&todo, todoId)
 
     if todo.ID == 0 {
-        response := newTodoFailResponse()
-        w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-        w.WriteHeader(404)
-        if err := json.NewEncoder(w).Encode(&response); err != nil {
-            panic(err)
-        }
+        RespondWithMessage(w, "No item was found with id " + todoId)
         return
     }
 
@@ -65,7 +57,14 @@ func TodoShow(w http.ResponseWriter, r *http.Request) {
 
 func TodoCreate(w http.ResponseWriter, r *http.Request) {
     db := storage.GetActiveDB()
-    todo :=  models.Todo{}
+    user := Authenticate(r, db)
+
+    if user.ID == 0 {
+        RespondWithMessage(w, "Unable to authenticate user.")
+        return
+    }
+
+    todo :=  models.Todo{ UserId: user.ID }
     body, err := ioutil.ReadAll(io.LimitReader(r.Body, 1048576))
     if err != nil {
         panic(err)
@@ -91,10 +90,22 @@ func TodoCreate(w http.ResponseWriter, r *http.Request) {
 
 func TodoUpdate(w http.ResponseWriter, r *http.Request) {
     db := storage.GetActiveDB()
+    user := Authenticate(r, db)
+
+    if user.ID == 0 {
+        RespondWithMessage(w, "Unable to authenticate user.")
+        return
+    }
+
     vars := mux.Vars(r)
     todoId := vars["todoId"]
     todo := models.Todo{}
-    db.Find(&todo, todoId)
+    db.Where("user_id = ?", user.ID).Find(&todo, todoId)
+
+    if todo.ID == 0 {
+        RespondWithMessage(w, "No items with id " + todoId + " found for this user.")
+        return
+    }
 
     body, err := ioutil.ReadAll(io.LimitReader(r.Body, 1048576))
     if err != nil {
@@ -121,10 +132,23 @@ func TodoUpdate(w http.ResponseWriter, r *http.Request) {
 
 func TodoDelete(w http.ResponseWriter, r *http.Request) {
     db := storage.GetActiveDB()
+    user := Authenticate(r, db)
+
+    if user.ID == 0 {
+        RespondWithMessage(w, "Unable to authenticate user.")
+        return
+    }
+
     vars := mux.Vars(r)
     todoId := vars["todoId"]
     todo := models.Todo{}
-    db.Find(&todo, todoId)
+    db.Where("user_id = ?", user.ID).Find(&todo, todoId)
+
+    if todo.ID == 0 {
+        RespondWithMessage(w, "No items with id " + todoId + " found for this user.")
+        return
+    }
+
     db.Delete(&todo, todoId)
 
     w.Header().Set("Content-Type", "application/json; charset=UTF-8")
